+++ Content.Server/_RMC14/Medical/Autodoc/AutodocSystem.cs
@@ -30,7 +30,6 @@ public sealed class AutodocSystem : SharedAutodocSystem
     [Dependency] private readonly AudioSystem _audio = default!;
     [Dependency] private readonly SharedBodySystem _body = default!;
     [Dependency] private readonly DamageableSystem _damageable = default!;
-    [Dependency] private readonly EntityLookupSystem _entityLookup = default!;
     [Dependency] private readonly MobStateSystem _mobState = default!;
     [Dependency] private readonly MobThresholdSystem _mobThreshold = default!;
     [Dependency] private readonly SharedPopupSystem _popup = default!;
@@ -42,7 +41,6 @@ public sealed class AutodocSystem : SharedAutodocSystem
     [Dependency] private readonly IGameTiming _timing = default!;
     [Dependency] private readonly UserInterfaceSystem _ui = default!;
 
-    private readonly HashSet<EntityUid> _intersecting = [];
     private readonly List<ProtoId<ReagentPrototype>> _reagentRemovalBuffer = [];
 
     private static readonly ProtoId<DamageGroupPrototype> BruteGroup = "Brute";
@@ -518,22 +516,6 @@ public sealed class AutodocSystem : SharedAutodocSystem
     public override void Update(float frameTime)
     {
         base.Update(frameTime);
-
-        var query = EntityQueryEnumerator<OutsideAutodocComponent>();
-        while (query.MoveNext(out var uid, out var comp))
-        {
-            if (comp.Chamber is not { } chamber)
-            {
-                RemCompDeferred<OutsideAutodocComponent>(uid);
-                continue;
-            }
-
-            _intersecting.Clear();
-            _entityLookup.GetEntitiesIntersecting(uid, _intersecting);
-
-            if (!_intersecting.Contains(chamber))
-                RemCompDeferred<OutsideAutodocComponent>(uid);
-        }
 
         var time = _timing.CurTime;
         var consoles = EntityQueryEnumerator<AutodocConsoleComponent>();
