+++ Content.Shared/_RMC14/Marines/HyperSleep/SharedHyperSleepChamberSystem.cs
@@ -1,7 +1,7 @@
+using Content.Shared._RMC14.Movement;
 using Content.Shared._RMC14.Xenonids;
 using Content.Shared.Movement.Events;
 using Robust.Shared.Containers;
-using Robust.Shared.Physics.Events;
 using Robust.Shared.Timing;
 
 namespace Content.Shared._RMC14.Marines.HyperSleep;
@@ -9,12 +9,10 @@ namespace Content.Shared._RMC14.Marines.HyperSleep;
 public abstract class SharedHyperSleepChamberSystem : EntitySystem
 {
     [Dependency] private readonly SharedContainerSystem _containers = default!;
-    [Dependency] private readonly EntityLookupSystem _entityLookup = default!;
+    [Dependency] private readonly RMCMovementSystem _rmcMovement = default!;
     [Dependency] private readonly IGameTiming _timing = default!;
 
     private EntityQuery<HyperSleepChamberComponent> _hyperSleepQuery;
-
-    private readonly HashSet<EntityUid> _intersecting = new();
 
     public override void Initialize()
     {
@@ -25,8 +23,6 @@ public abstract class SharedHyperSleepChamberSystem : EntitySystem
         SubscribeLocalEvent<HyperSleepChamberComponent, EntInsertedIntoContainerMessage>(OnInserted);
 
         SubscribeLocalEvent<InsideHyperSleepChamberComponent, MoveInputEvent>(OnMoveInput);
-
-        SubscribeLocalEvent<OutsideHyperSleepChamberComponent, PreventCollideEvent>(OnPreventCollide);
     }
 
     private void OnMapInit(Entity<HyperSleepChamberComponent> ent, ref MapInitEvent args)
@@ -61,16 +57,7 @@ public abstract class SharedHyperSleepChamberSystem : EntitySystem
             return;
 
         RemCompDeferred<InsideHyperSleepChamberComponent>(ent);
-
-        var outside = EnsureComp<OutsideHyperSleepChamberComponent>(ent);
-        outside.Chamber = chamber;
-        Dirty(ent, outside);
-    }
-
-    private void OnPreventCollide(Entity<OutsideHyperSleepChamberComponent> ent, ref PreventCollideEvent args)
-    {
-        if (ent.Comp.Chamber == args.OtherEntity)
-            args.Cancelled = true;
+        _rmcMovement.SuppressCollisionOnExit(ent, chamber);
     }
 
     public void EjectChamber(Entity<HyperSleepChamberComponent?> ent)
@@ -80,24 +67,5 @@ public abstract class SharedHyperSleepChamberSystem : EntitySystem
 
         if (_containers.TryGetContainer(ent, ent.Comp.ContainerId, out var container))
             _containers.EmptyContainer(container);
-    }
-
-    public override void Update(float frameTime)
-    {
-        var query = EntityQueryEnumerator<OutsideHyperSleepChamberComponent>();
-        while (query.MoveNext(out var uid, out var comp))
-        {
-            if (comp.Chamber is not { } chamber)
-            {
-                RemCompDeferred<OutsideHyperSleepChamberComponent>(uid);
-                continue;
-            }
-
-            _intersecting.Clear();
-            _entityLookup.GetEntitiesIntersecting(uid, _intersecting);
-
-            if (!_intersecting.Contains(chamber))
-                RemCompDeferred<OutsideHyperSleepChamberComponent>(uid);
-        }
     }
 }
