--- Content.Shared/_RMC14/Medical/Autodoc/OutsideAutodocComponent.cs
@@ -1,11 +1,0 @@
-using Robust.Shared.GameStates;
-
-namespace Content.Shared._RMC14.Medical.Autodoc;
-
-[RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
-[Access(typeof(SharedAutodocSystem))]
-public sealed partial class OutsideAutodocComponent : Component
-{
-    [DataField, AutoNetworkedField]
-    public EntityUid? Chamber;
-}
