--- Content.Shared/_RMC14/Medical/BodyScanner/OutsideBodyScannerComponent.cs
@@ -1,11 +1,0 @@
-using Robust.Shared.GameStates;
-
-namespace Content.Shared._RMC14.Medical.BodyScanner;
-
-[RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
-[Access(typeof(SharedBodyScannerSystem))]
-public sealed partial class OutsideBodyScannerComponent : Component
-{
-    [DataField, AutoNetworkedField]
-    public EntityUid? Chamber;
-}
