+++ Content.Shared/_RMC14/Medical/BodyScanner/SharedBodyScannerSystem.cs
@@ -1,3 +1,4 @@
+using Content.Shared._RMC14.Movement;
 using Content.Shared._RMC14.Storage;
 using Content.Shared._RMC14.Xenonids;
 using Content.Shared.Damage;
@@ -9,7 +10,6 @@ using Content.Shared.Stunnable;
 using Robust.Shared.Audio.Systems;
 using Robust.Shared.Containers;
 using Robust.Shared.Network;
-using Robust.Shared.Physics.Events;
 using Robust.Shared.Timing;
 
 namespace Content.Shared._RMC14.Medical.BodyScanner;
@@ -21,6 +21,7 @@ public abstract class SharedBodyScannerSystem : EntitySystem
     [Dependency] private readonly SharedContainerSystem _container = default!;
     [Dependency] private readonly INetManager _net = default!;
     [Dependency] private readonly SharedPopupSystem _popup = default!;
+    [Dependency] private readonly RMCMovementSystem _rmcMovement = default!;
     [Dependency] private readonly SharedStunSystem _stun = default!;
     [Dependency] private readonly IGameTiming _timing = default!;
     [Dependency] private readonly SharedTransformSystem _transform = default!;
@@ -39,8 +40,6 @@ public abstract class SharedBodyScannerSystem : EntitySystem
         SubscribeLocalEvent<BodyScannerConsoleComponent, ActivateInWorldEvent>(OnConsoleActivateInWorld);
 
         SubscribeLocalEvent<InsideBodyScannerComponent, MoveInputEvent>(OnInsideBodyScannerMoveInput);
-
-        SubscribeLocalEvent<OutsideBodyScannerComponent, PreventCollideEvent>(OnOutsideBodyScannerPreventCollide);
     }
 
     private void OnBodyScannerInit(Entity<BodyScannerComponent> scanner, ref ComponentInit args)
@@ -94,11 +93,7 @@ public abstract class SharedBodyScannerSystem : EntitySystem
         UpdateBodyScannerVisuals(scanner);
 
         if (!_timing.ApplyingState)
-        {
-            var inside = EnsureComp<InsideBodyScannerComponent>(args.Entity);
-            inside.BodyScanner = scanner;
-            Dirty(args.Entity, inside);
-        }
+            EnsureComp<InsideBodyScannerComponent>(args.Entity).Chamber = scanner;
     }
 
     private void OnBodyScannerEntRemoved(Entity<BodyScannerComponent> scanner, ref EntRemovedFromContainerMessage args)
@@ -114,10 +109,7 @@ public abstract class SharedBodyScannerSystem : EntitySystem
 
         UpdateBodyScannerVisuals(scanner);
         RemCompDeferred<InsideBodyScannerComponent>(args.Entity);
-
-        var outside = EnsureComp<OutsideBodyScannerComponent>(args.Entity);
-        outside.Chamber = scanner;
-        Dirty(args.Entity, outside);
+        _rmcMovement.SuppressCollisionOnExit(args.Entity, scanner.Owner);
     }
 
     private void OnBodyScannerInteractHand(Entity<BodyScannerComponent> scanner, ref InteractHandEvent args)
@@ -228,18 +220,12 @@ public abstract class SharedBodyScannerSystem : EntitySystem
         if (_timing.ApplyingState)
             return;
 
-        if (ent.Comp.BodyScanner is not { } scannerId)
+        if (ent.Comp.Chamber is not { } scannerId)
             return;
 
         if (!TryComp<BodyScannerComponent>(scannerId, out var scanner))
             return;
 
         EjectOccupant((scannerId, scanner), ent);
-    }
-
-    private void OnOutsideBodyScannerPreventCollide(Entity<OutsideBodyScannerComponent> ent, ref PreventCollideEvent args)
-    {
-        if (ent.Comp.Chamber == args.OtherEntity)
-            args.Cancelled = true;
     }
 }
