+++ Content.Shared/_RMC14/Movement/RMCOutsideChamberComponent.cs
@@ -1,10 +1,10 @@
 using Robust.Shared.GameStates;
 
-namespace Content.Shared._RMC14.Medical.Sleeper;
+namespace Content.Shared._RMC14.Movement;
 
 [RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
-[Access(typeof(SharedSleeperSystem))]
-public sealed partial class OutsideSleeperComponent : Component
+[Access(typeof(RMCMovementSystem))]
+public sealed partial class RMCOutsideChamberComponent : Component
 {
     [DataField, AutoNetworkedField]
     public EntityUid? Chamber;
